Update CMake build configuration to attempt to statically link against libovr
It's recommended that you decompress the Oculus SDK into external/OculusSDK, but it will attempt to find it through other means. If it cannot find it, it will default to the attempting to dynamically load the library.

It also removes OVR_CAPI and OVR_CAPI_Keys from the source tree.

This has been tested to work under OS X, although it should also work under Linux.